pyproject.toml
[project]
name = "imdb-to-justwatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
